requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blockbuster==1.5.24
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
coverage==7.9.1
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
forbiddenfruit==0.1.4
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-api==0.2.29
langgraph-checkpoint==2.0.26
langgraph-checkpoint-postgres==2.0.21
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.1.0
langgraph-sdk==0.1.69
langsmith==0.3.42
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.9.2
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
playwright==1.52.0
pluggy==1.6.0
propcache==0.3.1
psycopg==3.2.9
psycopg-pool==3.2.6
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2

watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
